Speed up calls to list, nextid, precommit and prepush by deferring html conversion to when reports are generated
