+++ parsing_test.go
@@ -20,7 +20,7 @@ body
 	assert.Nil(t, err)
 	assert.Equal(t, "REQ-TEST-SWL-1", r.ID)
 	assert.Equal(t, "title", r.Title)
-	assert.Equal(t, "<p>body</p>\n<p>body</p>\n", string(r.Body))
+	assert.Equal(t, "body\n\nbody\n", r.Body)
 	assert.Equal(t, "This is why.", r.Attributes["RATIONALE"])
 	assert.Equal(t, "exists", r.Attributes["ATTRIBUTE WHICH WILL NEVER EXIST"])
 	assert.Equal(t, []string{"REQ-TEST-SYS-1"}, r.ParentIds)
@@ -58,7 +58,7 @@ body
 	assert.Nil(t, err)
 	assert.Equal(t, "REQ-TEST-SWL-1", r.ID)
 	assert.Equal(t, "DELETED Some title", r.Title)
-	assert.Equal(t, "<p>body</p>\n", string(r.Body))
+	assert.Equal(t, "body\n", r.Body)
 	assert.Equal(t, "This is why.", r.Attributes["RATIONALE"])
 	assert.Equal(t, []string{"REQ-TEST-SYS-1"}, r.ParentIds)
 	assert.True(t, r.IsDeleted())
@@ -88,7 +88,7 @@ func TestParseReq_NoAttributes(t *testing.T) {
 	r, err := ParseReq(`REQ-TEST-SWL-1 title
 body`)
 	assert.Nil(t, err)
-	assert.Equal(t, "<p>body</p>\n", string(r.Body))
+	assert.Equal(t, "body", r.Body)
 }
 
 func TestParseReq_EmptyAttributesSection(t *testing.T) {
